fixing cloned targeting
