+++ src/constants/misc.js
@@ -1,5 +1,7 @@
 import { exchange as ExchangeConstants } from 'adex-constants'
 import { constants } from 'adex-models'
+import { translate } from 'services/translations/translations'
+
 const { BID_STATES, TIMEOUTS } = ExchangeConstants
 const { SignatureModes } = constants
 const AdTypes = constants.AdUnitsTypes.map(type => {
@@ -117,3 +119,42 @@ export const VALIDATOR_ANALYTICS_TIMEFRAMES = [
 	// { label: 'LABEL_MONTH', value: 'month' },
 	// { label: 'LABEL_YEAR', value: 'year' },
 ]
+
+const autocompleteLocationsSingleSelect = () => {
+	return constants.AllCountries.map(country => {
+		return {
+			label: country.name,
+			value: country.value,
+		}
+	})
+}
+
+const autocompleteGendersSingleSelect = () => {
+	return constants.Genders.map(gender => {
+		return {
+			label: translate(gender.split('_')[1]),
+			value: gender,
+		}
+	})
+}
+
+const autocompleteTagsSingleSelect = () => {
+	return constants.PredefinedTags.map(tag => {
+		return {
+			label: tag._id,
+			value: tag._id,
+		}
+	})
+}
+
+export const AcLocations = autocompleteLocationsSingleSelect()
+export const AcGenders = autocompleteGendersSingleSelect()
+export const AcTags = autocompleteTagsSingleSelect()
+
+// TODO: Extract in contstants and add labels
+export const SOURCES = {
+	locations: { src: AcLocations, collection: 'targeting' },
+	genders: { src: AcGenders, collection: 'targeting' },
+	tags: { src: AcTags, collection: 'targeting' },
+	custom: { src: [], collection: 'targeting' },
+}
